requirements.txt
fastapi==0.110.2
uvicorn==0.27.1
pydantic==2.5.3
python-multipart==0.0.9
pillow==10.3.0

